docker/docker build/requirements.txt
arango-orm==0.5.8
certifi==2019.11.28
chardet==3.0.4
Click==7.0
DateTime==4.3
Flask==1.1.1
Flask-HTTPAuth==4.5.0
future==0.18.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==2.16.0
mock==3.0.5
nose==1.3.7
python-arango==5.2.1
pytz==2019.3
requests==2.22.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
zope.interface>4.7.1